refactor: Move JwtUtil to Service directory
